GamingInstance/requirements.txt
certifi==2020.6.20
cycler==0.10.0
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.1
opencv-python==4.4.0.40
Pillow==7.2.0
pygame==1.9.6
pyparsing==2.4.7
python-dateutil==2.8.1
PyTurboJPEG==1.4.1
six==1.15.0
python-dotenv==0.14.0
coloredlogs
websocket-client
aiortc
psutil
boto3
botocore